Send trigger message to DLX if initial Cloud Tasks publication fails (#83)
* If a trigger/sequence start message fails to publish, send it to the DLX

Today, if e.g. an order fails because cloud tasks throws an error in the initial call, that order is lost (apart from logs). This way, we send it to the DLX instead, so that it can be resent.

In the error handler, determine if it is an initial trigger message by checking if the cloud task queue header is present.

* Make sinon dependency explicit to make depcheck happy